+++ app/upgrades/latest/upgrades.go
@@ -11,6 +11,7 @@ import (
 	"github.com/archway-network/archway/app/keepers"
 	"github.com/archway-network/archway/app/upgrades"
 	callbackTypes "github.com/archway-network/archway/x/callback/types"
+	cwerrorstypes "github.com/archway-network/archway/x/cwerrors/types"
 	"github.com/archway-network/archway/x/cwfees"
 	cwicatypes "github.com/archway-network/archway/x/cwica/types"
 )
@@ -37,6 +38,7 @@ var Upgrade = upgrades.Upgrade{
 		Added: []string{
 			callbackTypes.ModuleName,
 			cwfees.ModuleName,
+			cwerrorstypes.ModuleName,
 			icacontrollertypes.StoreKey,
 			cwicatypes.ModuleName,
 		},