+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Contains all the PRs that improved the code without changing the behaviors.
 - [#542](https://github.com/archway-network/archway/pull/542) - Adding x/cwica module
 - [#543](https://github.com/archway-network/archway/pull/543) - Bumping sdk to v0.47.9
 - [#544](https://github.com/archway-network/archway/pull/544) - Bumping sdk to v0.47.10
+- [#546](https://github.com/archway-network/archway/pull/546) - Adding x/cwerrors module
 
 ### Improvements